Refactor(tests): Refact action test for controller test.
