+++ src/test/resources/certificates/spring-certificate-test.properties
@@ -1,3 +1,3 @@
-server.ssl.trust-store=./build/certificates/blackduck-alert-test.truststore
+server.ssl.trust-store=./build/certs/blackduck-alert-test.truststore
 server.ssl.trust-store-password=changeit
 server.ssl.trust-store-type=JKS
